Restructure some pages
