+++ frontend/src/api/categories.js
@@ -1,0 +1,6 @@
+// categories API
+import { baseFetch } from "./baseApi";
+
+export async function getCategories() {
+  return baseFetch("/api/categories");
+}
