+++ frontend/src/pages/About/About.jsx
@@ -362,8 +362,8 @@
 "use client";
 
 import { useEffect } from "react";
-import Header from "../components/Header";
-import Footer from "../components/Footer";
+import Header from "../../components/Header";
+import Footer from "../../components/Footer";
 import { CheckCircle } from "lucide-react";
 
 /* ---------- helpers: full-width section ---------- */
