+++ frontend/src/pages/Courses/Components/Section.jsx
@@ -1,0 +1,18 @@
+const Section = ({ id, title, subtitle, action, children }) => (
+  <section id={id} className="w-screen overflow-x-hidden py-10 lg:py-14">
+    <div className="w-screen px-6 lg:px-12">
+      {(title || subtitle || action) && (
+        <div className="mb-6 flex items-end justify-between gap-4">
+          <div>
+            {title && <h2 className="text-2xl lg:text-3xl font-bold text-[#1d4ed8]">{title}</h2>}
+            {subtitle && <p className="text-slate-600 mt-2">{subtitle}</p>}
+          </div>
+          {action}
+        </div>
+      )}
+      {children}
+    </div>
+  </section>
+);
+
+export default Section;
